requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
flask-migrate==4.0.5
python-dotenv==0.19.0
Flask-Bootstrap
flask-wtf
email-validator
weasyprint==54.3
Flask-Login==0.6.3
email-validator==1.3.1
reportlab>=3.6.0